requirements/general_requirments.txt
numpy
matplotlib
sklearn
segmentation-models
tensorflow-addons
scikit-image
tqdm
Pillow
scipy
seaborn